pyproject.toml
[project]
name = "test-cline"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.115.12",
    "google-genai>=1.12.1",
    "pytest>=8.3.5",
    "pytest-mock>=3.14.0",
    "requests>=2.32.3",
    "uvicorn>=0.34.2",
]
